requirements.txt
openai>=1.6.1
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
streamlit>=1.35.0